Organized code.
Moved GameWorld and GameMap to separate files.
